refactor(calendar): extract getTimeOptions to shared src/lib/timeOptions
